// backend/src/modules/matchmaker/domain/enums/match-status.enum.ts
export enum MatchStatus {
  OFFERED = 'OFFERED',
  ACCEPTED = 'ACCEPTED',
  EN_ROUTE = 'EN_ROUTE',
  ARRIVED = 'ARRIVED',
  COMPLETED = 'COMPLETED',
  FAILED_SPOT_TAKEN = 'FAILED_SPOT_TAKEN',
  CANCELLED_SEARCHER = 'CANCELLED_SEARCHER',
  CANCELLED_LEAVER = 'CANCELLED_LEAVER',
  TIMEOUT = 'TIMEOUT',
  DECLINED = 'DECLINED',
}

export type MatchType = 'REAL_TIME_P2P' | 'PROBABILISTIC_DB';